# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SubscribeRequest {
  /** @maxLength 320 */
  email: string;
  /** @maxItems 10 */
  layers: string[];
}

export interface SubscribeResponse {
  ok: boolean;
  message: string;
}

export interface ErrorResponse {
  ok: boolean;
  message: string;
}

export type AssistantMessageRole = typeof AssistantMessageRole[keyof typeof AssistantMessageRole];


export const AssistantMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AssistantMessage {
  role: AssistantMessageRole;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
}

export interface AssistantConversation {
  /**
     * @minItems 1
     * @maxItems 30
     */
  messages: AssistantMessage[];
}

export interface AssistantReply {
  reply: string;
}