# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: subscribe
    description: Specification update feed subscriptions
  - name: assistant
    description: AI assistant grounded in the whitepaper
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /subscribe:
    post:
      operationId: subscribeToUpdates
      tags: [subscribe]
      summary: Subscribe to specification update feed
      description: Registers an email address for RFC specification updates via Resend.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeRequest"
      responses:
        "200":
          description: Email registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscribeResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server misconfiguration (e.g. email provider not configured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Upstream email provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assistant/ask:
    post:
      operationId: askWhitepaper
      tags: [assistant]
      summary: Ask the whitepaper assistant a question
      description: >-
        Answers a reader's question using only the AI-Native Office whitepaper as
        grounding. Accepts the full conversation so far (prior turns plus the new
        question) and returns the assistant's reply. Questions outside the
        whitepaper's scope receive an in-character refusal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantConversation"
      responses:
        "200":
          description: Assistant reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantReply"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server misconfiguration (e.g. AI provider not configured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Upstream AI provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SubscribeRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        layers:
          type: array
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 200
      required:
        - email
        - layers
    SubscribeResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
      required:
        - ok
        - message
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
      required:
        - ok
        - message
    AssistantMessage:
      type: object
      properties:
        role:
          type: string
          enum:
            - user
            - assistant
        content:
          type: string
          minLength: 1
          maxLength: 4000
      required:
        - role
        - content
    AssistantConversation:
      type: object
      properties:
        messages:
          type: array
          minItems: 1
          maxItems: 30
          items:
            $ref: "#/components/schemas/AssistantMessage"
      required:
        - messages
    AssistantReply:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply