# lib/api-zod/src/generated/types/assistantConversation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantMessage } from './assistantMessage';

export interface AssistantConversation {
  /**
     * @minItems 1
     * @maxItems 30
     */
  messages: AssistantMessage[];
}