# lib/api-zod/src/generated/types/assistantMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantMessageRole } from './assistantMessageRole';

export interface AssistantMessage {
  role: AssistantMessageRole;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
}